Replaced unecessary "z" index dimension with "depth" coordinate
Currently `ctd_bincast()` has `z` as its dimension, but the values of this are just integer indices. This is bound to cause user errors because if `zmin` is small and `dz=1.` then `z` looks like it might be a depth variable. I don't really see a purpose for a variable that is just the indices of the dimensions, so I removed it and just replaced it with the `depth` coordinate.