Clarified that CTD just reads data in physical units
I think it is important to be clear about which data processing steps happen when in the pipeline. My understanding is all of the functions in `io.py` just extract the data and metadata from the hex files and apply Seabird's formulas for converting the engineering units to physical units. I can't imagine users will want to fiddle with this step as much so I would refrain from calling this step "processing" and only use the term "processing" when referring to all of the post-processing after physical data is extracted.