Embraced xarray by making parameters attributes
- Made all hard-coded parameters in proc.py attributes of the CTD datasets instead. One needs to be a bit careful to make sure these attributes are carried through at some stages where a new dataset is made from scratch and filled with re-processed ctd (e.g. in phase_correct).

- Renamed "data" variable to "ds" to emphasize that the arguments should be `xr.Dataset` types. This can get confusing in some places because we had used commands like `data['t1'].data` to access the underlying data numpy arrays in a dataset. This new naming convention should make it more clear what is expected to be a dataset and what is expected to just be a numpy array (i.e. the underlying data in the dataset, but possibly just a subset of the full dataset).

Note: A more efficient way of doing this would be to then keep the same xr.Dataset instance for the entire processing, saving on memory allocations. Instead, however, we frequently make shallow and return shallow copies with some modifications and overwrite the dataset. In a couple of cases, we actually completely start a dataset from scratch and fill it will data. It probably doesn't matter since these operations still run in almost-real-time for a single profile (so within a couple minutes for a whole cruise-worth of data), it would be prohibitive for very large datasets (or possibly if trying to process high-frequency and long-record mooring / float data with these same functions).

- Made function names more concise and more reflective of what they actually do.

- Replaced repeated function calls (e.g. a processing step applied to both t1 and t2 or to all variables) with for loops. With consistent naming of parameters in ds.attrs this allows a different parameter value to be passed as an argument for each variable. For example, for the spike threshold parameter for temperature and salinity, 'spike_thresh_t' or 'spike_thresh_s' respectively, these parameters can be passed with a for loop by:

```python
for v in ['t1', 's1']:
	despike(ds[v], ds.attrs[f'spike_thresh_{v[0]}'])
```

- Cleaned up some deprecation warnings (on 3.9, hopefully not breaking for 3.6-3.8)