Fixed some breaking changes that passed tests
I had run the unit tests but not actually tried fully processing a profile with run_all. It made me realize that we need a end-to-end unit test, even if it is just for an artificial ctd dataset or something, just to make sure all of the individual components integrate properly.

There's also no reason to stop with `run_all()`. We could have a function `full_process` which includes reading the hex file, running all the processing, saving as netcdf files, etc at in one function.