Fixed test for unique_arrays
The test was broken for unique arrays because I changed what its implementation slightly. The old implementation was never actually necessary and relied on a very clunky (and deprecated) ragged array data structure. Instead it now just has to append and/or stack lists together. The new test now matches that implementation.